pyproject.toml
[project]
name = "wandb-docs-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai-agents",
    "openai>=1.74.0",
    "wandb",
    "weave>=0.51.43",
    "python-dotenv",
    "aiohttp",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.1",
    "requests>=2.32.3",
    "pydantic>=2.11.3",
    "simple-parsing>=0.1.7",
    "httpx>=0.28.1",
    "typing-extensions>=4.13.2",
]

[project.optional-dependencies]
tests = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]

[tool.setuptools.packages.find]
where = ["src"]
